requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.5.14
dill==0.3.7
editdistpy==0.1.3
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
joblib==1.3.2
langchain==0.0.258
langsmith==0.0.83
marshmallow==3.20.2
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.5
numexpr==2.8.8
numpy==1.26.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.2
pandas==2.2.0
pydantic==1.10.14
pyfunctional==1.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.2.2
scipy==1.11.4
six==1.16.0
SQLAlchemy==2.0.25
symspellpy==6.7.7
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
trieregex==1.0.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
xgboost==1.7.5
yarl==1.9.4
